refactor: separate Main and App from the home component
