+++ src/App.tsx
@@ -1,0 +1,15 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import { Home } from "./Home"
+
+
+const queryClient = new QueryClient();
+
+function App () {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <Home />
+    </QueryClientProvider>
+  );
+}
+
+export default App
